+++ codegen/keyword_generation/data_model/__init__.py
@@ -1,4 +1,7 @@
 
+import copy
+import json
+import os
 import typing
 
 KWDM_INSTANCE = None
@@ -37,3 +40,46 @@ def get_aliased_by(keyword: str):
 
 def is_aliased(keyword: str):
     return keyword in ALIAS_TO_KWD.keys()
+
+def _load_manifest(filename) -> typing.Dict:
+    with open(filename) as f:
+        manifest = json.load(f)
+    return manifest
+
+class AdditionalCards:
+    def __init__(self, filename):
+        with open(filename) as f:
+            self._cards = json.load(f)
+
+    def __getitem__(self, name):
+        """return a copy of the additional card, since the client may mutate it."""
+        return copy.deepcopy(self._cards[name])
+
+class KWDM:
+    def __init__(self, filename):
+        with open(filename, encoding="utf-8") as f:
+            self._data: typing.Dict = json.load(f)
+
+    def get_keywords_list(self) -> typing.List[str]:
+        return list(self._data.keys())
+
+    def get_keyword_data_dict(self, name: str) -> typing.Dict:
+        return copy.deepcopy(self[name])
+
+    def __getitem__(self, name: str) -> typing.Dict:
+        return self._data[name]
+
+def load(this_folder: str, kwd_file: str, manifest: str, additional_cards: str):
+    global KWDM_INSTANCE, MANIFEST, ADDITIONAL_CARDS
+    if kwd_file == "":
+        KWDM_INSTANCE = KWDM(os.path.join(this_folder, "kwd.json"))
+    else:
+        KWDM_INSTANCE = KWDM(kwd_file)
+    if manifest == "":
+        MANIFEST = _load_manifest(this_folder / "manifest.json")
+    else:
+        MANIFEST = _load_manifest(manifest)
+    if additional_cards == "":
+        ADDITIONAL_CARDS = AdditionalCards(this_folder / "additional-cards.json")
+    else:
+        ADDITIONAL_CARDS = AdditionalCards(additional_cards)
